chore: organize plugin folder
