+++ engine/entity/component/AudioWavStreamComponent.h
@@ -2,7 +2,7 @@
 
 #include <string>
 #include <soloud_wavstream.h>
-#include <plugin/Audio.h>
+#include <plugin/audio/Audio.h>
 #include <resource/BinaryResource.h>
 #include "TransformComponent.h"
