+++ engine/plugin/discord/Discord.cpp
@@ -5,7 +5,6 @@
 #include <config/ConEntry.h>
 #include <core/Logger.h>
 #include <i18n/TranslationManager.h>
-#include "Plugin.h"
 
 using namespace chira;
