+++ engine/plugin/steam/CMakeLists.txt
@@ -1,0 +1,5 @@
+list(APPEND CHIRA_ENGINE_HEADERS
+        ${CMAKE_CURRENT_LIST_DIR}/Steam.h)
+
+list(APPEND CHIRA_ENGINE_SOURCES
+        ${CMAKE_CURRENT_LIST_DIR}/Steam.cpp)
