+++ engine/plugin/steam/Steam.cpp
@@ -4,9 +4,9 @@
 #include <libloader.hpp>
 #include <config/ConEntry.h>
 #include <core/Logger.h>
+#include <plugin/Plugin.h>
 #include <resource/provider/FilesystemResourceProvider.h>
 #include <utility/String.h>
-#include "Plugin.h"
 
 using namespace chira;
 using namespace libloader;
